Weather Cards - Bug fix
When choosing a location for the 2nd+ time, the card would almost always load in the location name first and then the weather stats because of varying API response times. Fixed it by nesting fetch calls and if all goes well, set locations and weather stats at the same time.

Did not add functionality to handle if there is an error in the API fetching process.